+++ tests/test_needflow.py
@@ -47,6 +47,9 @@ def test_doc_build_html(test_app):
     assert "STORY_2 [[../index.html#STORY_2]]" in page_html
     assert "STORY_2.another_one [[../index.html#STORY_2.another_one]]" in page_html
 
+    empty_needflow_with_debug = Path(app.outdir, "empty_needflow_with_debug.html").read_text()
+    assert "No needs passed the filters" in empty_needflow_with_debug
+
 
 @pytest.mark.parametrize(
     "test_app", [{"buildername": "html", "srcdir": "doc_test/doc_needflow_incl_child_needs"}], indirect=True
@@ -61,18 +64,82 @@ def test_doc_build_needflow_incl_child_needs(test_app):
     # plantuml shall not return any warnings:
     assert "WARNING: error while running plantuml" not in warnings
 
-    html = Path(app.outdir, "index.html").read_text()
-    assert html
-    assert "@startuml" in html
-    assert "[[../index.html#STORY_1]]" in html
-    assert "[[../index.html#STORY_1.1]]" in html
-    assert "[[../index.html#STORY_1.2]]" in html
-    assert "[[../index.html#STORY_2]]" in html
-    assert "[[../index.html#STORY_2.3]]" in html
-    assert "[[../index.html#SPEC_1]]" in html
-    assert "[[../index.html#SPEC_2]]" in html
-    assert "[[../index.html#SPEC_3]]" in html
-    assert "[[../index.html#SPEC_4]]" in html
-    assert "[[../index.html#STORY_3]]" in html
-    assert "[[../index.html#SPEC_5]]" in html
-    assert "@enduml" in html
+    index_html = Path(app.outdir, "index.html").read_text()
+    assert index_html
+    assert index_html.count("@startuml") == 1
+    assert index_html.count("[[../index.html#STORY_1]]") == 2
+    assert index_html.count("[[../index.html#STORY_1.1]]") == 2
+    assert index_html.count("[[../index.html#STORY_1.2]]") == 2
+    assert index_html.count("[[../index.html#STORY_2]]") == 2
+    assert index_html.count("[[../index.html#STORY_2.3]]") == 2
+    assert index_html.count("[[../index.html#SPEC_1]]") == 2
+    assert index_html.count("[[../index.html#SPEC_2]]") == 2
+    assert index_html.count("[[../index.html#SPEC_3]]") == 2
+    assert index_html.count("[[../index.html#SPEC_4]]") == 2
+    assert index_html.count("[[../index.html#STORY_3]]") == 2
+    assert index_html.count("[[../index.html#SPEC_5]]") == 2
+    assert index_html.count("@enduml") == 1
+
+    single_parent_need_filer_html = Path(app.outdir, "single_parent_need_filer.html").read_text()
+    assert single_parent_need_filer_html
+    assert single_parent_need_filer_html.count("@startuml") == 1
+    assert single_parent_need_filer_html.count("[[../index.html#STORY_3]]") == 2
+    assert single_parent_need_filer_html.count("@enduml") == 1
+    assert "[[../index.html#STORY_1]]" not in single_parent_need_filer_html
+    assert "[[../index.html#STORY_1.1]]" not in single_parent_need_filer_html
+    assert "[[../index.html#STORY_1.2]]" not in single_parent_need_filer_html
+    assert "[[../index.html#STORY_2]]" not in single_parent_need_filer_html
+    assert "[[../index.html#STORY_2.3]]" not in single_parent_need_filer_html
+    assert "[[../index.html#SPEC_1]]" not in single_parent_need_filer_html
+    assert "[[../index.html#SPEC_2]]" not in single_parent_need_filer_html
+    assert "[[../index.html#SPEC_3]]" not in single_parent_need_filer_html
+    assert "[[../index.html#SPEC_4]]" not in single_parent_need_filer_html
+    assert "[[../index.html#SPEC_5]]" not in single_parent_need_filer_html
+
+    single_child_with_child_need_filter_html = Path(app.outdir, "single_child_with_child_need_filter.html").read_text()
+    assert single_child_with_child_need_filter_html
+    assert single_child_with_child_need_filter_html.count("@startuml") == 1
+    assert single_child_with_child_need_filter_html.count("[[../index.html#STORY_2]]") == 2
+    assert single_child_with_child_need_filter_html.count("@enduml") == 1
+    assert "[[../index.html#STORY_1]]" not in single_child_with_child_need_filter_html
+    assert "[[../index.html#STORY_1.1]]" not in single_child_with_child_need_filter_html
+    assert "[[../index.html#STORY_1.2]]" not in single_child_with_child_need_filter_html
+    assert "[[../index.html#STORY_2.3]]" not in single_child_with_child_need_filter_html
+    assert "[[../index.html#SPEC_1]]" not in single_child_with_child_need_filter_html
+    assert "[[../index.html#SPEC_2]]" not in single_child_with_child_need_filter_html
+    assert "[[../index.html#SPEC_3]]" not in single_child_with_child_need_filter_html
+    assert "[[../index.html#SPEC_4]]" not in single_child_with_child_need_filter_html
+    assert "[[../index.html#STORY_3]]" not in single_child_with_child_need_filter_html
+    assert "[[../index.html#SPEC_5]]" not in single_child_with_child_need_filter_html
+
+    single_child_need_filter_html = Path(app.outdir, "single_child_need_filter.html").read_text()
+    assert single_child_need_filter_html
+    assert single_child_need_filter_html.count("@startuml") == 1
+    assert single_child_need_filter_html.count("[[../index.html#SPEC_1]]") == 2
+    assert single_child_need_filter_html.count("@enduml") == 1
+    assert "[[../index.html#STORY_1]]" not in single_child_need_filter_html
+    assert "[[../index.html#STORY_1.1]]" not in single_child_need_filter_html
+    assert "[[../index.html#STORY_1.2]]" not in single_child_need_filter_html
+    assert "[[../index.html#STORY_2]]" not in single_child_need_filter_html
+    assert "[[../index.html#STORY_2.3]]" not in single_child_need_filter_html
+    assert "[[../index.html#SPEC_2]]" not in single_child_need_filter_html
+    assert "[[../index.html#SPEC_3]]" not in single_child_need_filter_html
+    assert "[[../index.html#SPEC_4]]" not in single_child_need_filter_html
+    assert "[[../index.html#STORY_3]]" not in single_child_need_filter_html
+    assert "[[../index.html#SPEC_5]]" not in single_child_need_filter_html
+
+    grandy_and_child_html = Path(app.outdir, "grandy_and_child.html").read_text()
+    assert grandy_and_child_html
+    assert grandy_and_child_html.count("@startuml") == 1
+    assert grandy_and_child_html.count("[[../index.html#STORY_1]]") == 2
+    assert grandy_and_child_html.count("[[../index.html#SPEC_1]]") == 2
+    assert grandy_and_child_html.count("[[../index.html#SPEC_2]]") == 2
+    assert grandy_and_child_html.count("@enduml") == 1
+    assert "[[../index.html#STORY_1.1]]" not in grandy_and_child_html
+    assert "[[../index.html#STORY_1.2]]" not in grandy_and_child_html
+    assert "[[../index.html#STORY_2]]" not in grandy_and_child_html
+    assert "[[../index.html#STORY_2.3]]" not in grandy_and_child_html
+    assert "[[../index.html#SPEC_3]]" not in grandy_and_child_html
+    assert "[[../index.html#SPEC_4]]" not in grandy_and_child_html
+    assert "[[../index.html#STORY_3]]" not in grandy_and_child_html
+    assert "[[../index.html#SPEC_5]]" not in grandy_and_child_html